requirements.txt
mido==1.3.3
numpy==2.4.4
packaging==26.0
pyfluidsynth==1.3.4
python-rtmidi==1.5.8
